Renamed interaction db file
